feat: Ajusta retorno de requisições padronizando as repostas e ocultando dados sensíveis

+++ src/post/post.service.ts
@@ -1,33 +1,43 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { PrismaService } from 'src/database/prisma.service';
 
 @Injectable()
 export class PostService {
-  constructor(private readonly prisma: PrismaService, ) {}
+  constructor(private readonly prisma: PrismaService) {}
 
   async create(createPostDto: CreatePostDto, userId: number) {
+    const user = await this.prisma.user.findUnique({ where: { id: userId } });
+
+    if (!user) {
+      throw new NotFoundException(`User with id ${userId} not found`);
+    }
+
     return await this.prisma.post.create({
-     data: {
-      ...createPostDto,
-      user: {
-        connect: { id: userId },
+      data: {
+        ...createPostDto,
+        user: {
+          connect: { id: userId },
+        },
       },
-    },
     });
   }
 
   async findAll() {
     return await this.prisma.post.findMany({
-      select:{
-        answers:true,
-        user:{
-          select:{
+      select: {
+        id: true,
+        title: true,
+        content: true,
+        createdAt: true,
+        user: {
+          select: {
             name: true,
             email: true,
-          }
-        }
+          },
+        },
+        answers: true,
       },
     });
   }
@@ -37,18 +47,43 @@ export class PostService {
   }
 
   async update(id: number, updatePostDto: UpdatePostDto) {
+    const post = await this.prisma.post.findUnique({ where: { id } });
+
+    if (!post) {
+      throw new NotFoundException(`Post with ${id} not found`);
+    }
+
     return await this.prisma.post.update({
-      where: {id},
+      where: { id },
       data: updatePostDto,
+      select: {
+        id: true,
+        title: true,
+        content: true,
+        createdAt: true,
+        updatedAt: true,
+        user:{
+          select:{
+            email:true,
+            name:true,
+          }
+
+        }
+      },
     });
   }
 
   async remove(id: number) {
-    
-    await this.prisma.answers.deleteMany({
-      where:{postId: id}
-    })
+    const post = await this.prisma.post.findUnique({ where: { id } });
 
-    return await this.prisma.post.delete({where: {id}});
+    if (!post) {
+      throw new NotFoundException(`Post with ${id} not found`);
+    }
+
+    await this.prisma.answers.deleteMany({
+      where: { postId: id },
+    });
+
+    return await this.prisma.post.delete({ where: { id } });
   }
 }
